Add attributes for inference of static and dynamic aliases
These attributes are used for inference. The fallback case (when no attributes)
is always to infer a static alias (will be changed latter when introducing
dynamic aliases of arbitrary paths).

- Update test cases that display aliases as module types in errors. Static
aliases as signature items are displayed as ambiguous aliases `module X = P`,
while dynamic ones are displayed with an attribute `module X = P
[@@dynamic_alias]`.