Prevent inference of present static aliases via avoidance
Through avoidance, present static aliases could be inferred. This commit removes
the logic of [nondep_mty_with_presence], instead relying on a match on the
result of avoidance to decide on presence. Therefore, avoidance can still infer
static aliases, but the presence flag always correspond to the static/dynamic
type of alias.